bold colors

+++ src/ls/execute.cpp
@@ -95,8 +95,8 @@ void print_files(const std::vector<File>& files, int flags) {
 	struct winsize w;
 	ioctl(0, TIOCGWINSZ, &w);
 	int max_width = w.ws_col;
-	size_t n_rows = 1;
-	std::vector<size_t> column_width(files.size()+1);
+	unsigned n_rows = 1;
+	std::vector<unsigned> column_width(files.size()+1);
 	while (n_rows < files.size()) {
 		int column = -1;
 		for (size_t i = 0; i < files.size(); ++i) {
